fix login and sign_up
